feat: extracting handlers
